refactor(outfitter): extract upgrade output and report modules (OS-433)
